+++ markup/favorites.html
@@ -67,7 +67,7 @@
                     </div>
                     <div class="favorites__image-wrapper place-card__image-wrapper">
                       <a href="#">
-                        <img class="place-card__image" src="img/apartment-small-03.jpg" width="150" height="110" alt="Place image">
+                        <img class="place-card__image" src="img/apartment-small-03.jpg" width="150" height="110" alt="Place previewImage">
                       </a>
                     </div>
                     <div class="favorites__card-info place-card__info">
@@ -99,7 +99,7 @@
                   <article class="favorites__card place-card">
                     <div class="favorites__image-wrapper place-card__image-wrapper">
                       <a href="#">
-                        <img class="place-card__image" src="img/room-small.jpg" width="150" height="110" alt="Place image">
+                        <img class="place-card__image" src="img/room-small.jpg" width="150" height="110" alt="Place previewImage">
                       </a>
                     </div>
                     <div class="favorites__card-info place-card__info">
@@ -142,7 +142,7 @@
                   <article class="favorites__card place-card">
                     <div class="favorites__image-wrapper place-card__image-wrapper">
                       <a href="#">
-                        <img class="place-card__image" src="img/apartment-small-04.jpg" width="150" height="110" alt="Place image">
+                        <img class="place-card__image" src="img/apartment-small-04.jpg" width="150" height="110" alt="Place previewImage">
                       </a>
                     </div>
                     <div class="favorites__card-info place-card__info">
