+++ markup/login.html
@@ -37,11 +37,11 @@
             <form class="login__form form" action="#" method="post">
               <div class="login__input-wrapper form__input-wrapper">
                 <label class="visually-hidden">E-mail</label>
-                <input class="login__input form__input" type="email" name="email" placeholder="Email" required="">
+                <input class="login__input form__input" type="email" name="email" placeholder="Email" required>
               </div>
               <div class="login__input-wrapper form__input-wrapper">
                 <label class="visually-hidden">Password</label>
-                <input class="login__input form__input" type="password" name="password" placeholder="Password" required="">
+                <input class="login__input form__input" type="password" name="password" placeholder="Password" required>
               </div>
               <button class="login__submit form__submit button" type="submit">Sign in</button>
             </form>
