+++ markup/main.html
@@ -112,7 +112,7 @@
                   </div>
                   <div class="cities__image-wrapper place-card__image-wrapper">
                     <a href="#">
-                      <img class="place-card__image" src="img/apartment-01.jpg" width="260" height="200" alt="Place image">
+                      <img class="place-card__image" src="img/apartment-01.jpg" width="260" height="200" alt="Place previewImage">
                     </a>
                   </div>
                   <div class="place-card__info">
@@ -144,7 +144,7 @@
                 <article class="cities__card place-card">
                   <div class="cities__image-wrapper place-card__image-wrapper">
                     <a href="#">
-                      <img class="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place image">
+                      <img class="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place previewImage">
                     </a>
                   </div>
                   <div class="place-card__info">
@@ -176,7 +176,7 @@
                 <article class="cities__card place-card">
                   <div class="cities__image-wrapper place-card__image-wrapper">
                     <a href="#">
-                      <img class="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place image">
+                      <img class="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place previewImage">
                     </a>
                   </div>
                   <div class="place-card__info">
@@ -211,7 +211,7 @@
                   </div>
                   <div class="cities__image-wrapper place-card__image-wrapper">
                     <a href="#">
-                      <img class="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place image">
+                      <img class="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place previewImage">
                     </a>
                   </div>
                   <div class="place-card__info">
@@ -243,7 +243,7 @@
                 <article class="cities__card place-card">
                   <div class="cities__image-wrapper place-card__image-wrapper">
                     <a href="#">
-                      <img class="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place image">
+                      <img class="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place previewImage">
                     </a>
                   </div>
                   <div class="place-card__info">
