+++ markup/offer-not-logged.html
@@ -198,7 +198,7 @@
               <article class="near-places__card place-card">
                 <div class="near-places__image-wrapper place-card__image-wrapper">
                   <a href="#">
-                    <img class="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place image">
+                    <img class="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place previewImage">
                   </a>
                 </div>
                 <div class="place-card__info">
@@ -230,7 +230,7 @@
               <article class="near-places__card place-card">
                 <div class="near-places__image-wrapper place-card__image-wrapper">
                   <a href="#">
-                    <img class="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place image">
+                    <img class="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place previewImage">
                   </a>
                 </div>
                 <div class="place-card__info">
@@ -265,7 +265,7 @@
                 </div>
                 <div class="near-places__image-wrapper place-card__image-wrapper">
                   <a href="#">
-                    <img class="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place image">
+                    <img class="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place previewImage">
                   </a>
                 </div>
                 <div class="place-card__info">
