+++ markup/offer.html
@@ -250,7 +250,7 @@
               <article class="near-places__card place-card">
                 <div class="near-places__image-wrapper place-card__image-wrapper">
                   <a href="#">
-                    <img class="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place image">
+                    <img class="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place previewImage">
                   </a>
                 </div>
                 <div class="place-card__info">
@@ -282,7 +282,7 @@
               <article class="near-places__card place-card">
                 <div class="near-places__image-wrapper place-card__image-wrapper">
                   <a href="#">
-                    <img class="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place image">
+                    <img class="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place previewImage">
                   </a>
                 </div>
                 <div class="place-card__info">
@@ -317,7 +317,7 @@
                 </div>
                 <div class="near-places__image-wrapper place-card__image-wrapper">
                   <a href="#">
-                    <img class="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place image">
+                    <img class="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place previewImage">
                   </a>
                 </div>
                 <div class="place-card__info">
