--- src/components/NotFound.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const NotFound = () => (
-  <div style={{paddingLeft: 20}}>
-    <h1>404 Not Found</h1>
-  </div>
-);
-
-export default NotFound;
