+++ src/pages/NotFound.tsx
@@ -1,0 +1,9 @@
+function NotFound(): JSX.Element {
+  return (
+    <div style={{paddingLeft: 20}}>
+      <h1>404 Not Found</h1>
+    </div>
+  );
+}
+
+export default NotFound;
